Exception handling delete

+++ src/main/java/com/paulohenrique/course/services/UserService.java
@@ -2,8 +2,11 @@ package com.paulohenrique.course.services;
 
 import com.paulohenrique.course.entities.User;
 import com.paulohenrique.course.repositories.UserRepository;
+import com.paulohenrique.course.services.exceptions.DatabaseException;
 import com.paulohenrique.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -27,7 +30,15 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        try {
+            if(userRepository.existsById(id)){
+                userRepository.deleteById(id);
+            }else {
+                throw new ResourceNotFoundException(id);
+            }
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user){
